online message redis has been clustered from link https://www.javacodegeeks.com/2015/09/redis-clustering.html#RedisClusteringNut and code has been changed to use JedisCluster class

+++ engine/src/main/java/com/jfms/engine/service/biz/remote/GroupConverter.java
@@ -15,7 +15,7 @@ public class GroupConverter {
                 jfmsClientGroupCreationMessage.getDisplayName(),
                 jfmsClientGroupCreationMessage.getCreator(),
                 jfmsClientGroupCreationMessage.getJfmsGroupMemberMap(),
-                jfmsClientGroupCreationMessage.getType()
+                jfmsClientGroupCreationMessage.getType().byteValue()
         );
     }
 
@@ -24,7 +24,7 @@ public class GroupConverter {
                 jfmsClientGroupInfoEditMessage.getDisplayName(),
                 jfmsClientGroupInfoEditMessage.getCreator(),
                 jfmsClientGroupInfoEditMessage.getJfmsGroupMemberMap(),
-                jfmsClientGroupInfoEditMessage.getType()
+                jfmsClientGroupInfoEditMessage.getType().byteValue()
         );
     }
